[Maps] Support query-time runtime fields (#95701)
