+++ x-pack/test/functional/apps/maps/joins.js
@@ -39,7 +39,7 @@ export default function ({ getPageObjects, getService }) {
 
     it('should re-fetch join with refresh timer', async () => {
       async function getRequestTimestamp() {
-        await PageObjects.maps.openInspectorRequest('meta_for_geo_shapes*.shape_name');
+        await PageObjects.maps.openInspectorRequest('meta_for_geo_shapes*.runtime_shape_name');
         const requestStats = await inspector.getTableData();
         const requestTimestamp = PageObjects.maps.getInspectorStatRowHit(
           requestStats,
@@ -121,7 +121,9 @@ export default function ({ getPageObjects, getService }) {
       });
 
       it('should not apply query to source and apply query to join', async () => {
-        const joinResponse = await PageObjects.maps.getResponse('meta_for_geo_shapes*.shape_name');
+        const joinResponse = await PageObjects.maps.getResponse(
+          'meta_for_geo_shapes*.runtime_shape_name'
+        );
         expect(joinResponse.aggregations.join.buckets.length).to.equal(2);
       });
     });
@@ -136,7 +138,9 @@ export default function ({ getPageObjects, getService }) {
       });
 
       it('should apply query to join request', async () => {
-        const joinResponse = await PageObjects.maps.getResponse('meta_for_geo_shapes*.shape_name');
+        const joinResponse = await PageObjects.maps.getResponse(
+          'meta_for_geo_shapes*.runtime_shape_name'
+        );
         expect(joinResponse.aggregations.join.buckets.length).to.equal(1);
       });
